[turnRestrictedPath] moving rule.h to cpp_common
