+++ include/trsp/pgr_trspHandler.h
@@ -41,7 +41,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 #include "cpp_common/basePath_SSEC.hpp"
 #include "trsp/edgeInfo.h"
-#include "trsp/rule.h"
+#include "cpp_common/rule.h"
 
 namespace pgrouting {
 namespace trsp {
