+++ src/cpp_common/rule.cpp
@@ -23,7 +23,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#include "trsp/rule.h"
+#include "cpp_common/rule.h"
 
 #include <vector>
 #include <algorithm>
