+++ src/ksp/turnRestrictedPath_driver.cpp
@@ -37,7 +37,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "cpp_common/pgr_alloc.hpp"
 #include "cpp_common/pgr_assert.h"
 
-#include "trsp/rule.h"
+#include "cpp_common/rule.h"
 
 #include "cpp_common/basePath_SSEC.hpp"
 #include "dijkstraTR/pgr_dijkstraTR.hpp"
