+++ src/trsp/trsp_driver.cpp
@@ -31,7 +31,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <sstream>
 #include <algorithm>
 #include "trsp/pgr_trspHandler.h"
-#include "trsp/rule.h"
+#include "cpp_common/rule.h"
 #include "cpp_common/pgr_assert.h"
 #include "cpp_common/pgr_alloc.hpp"
